Jitter below 1 ns. Control panel in progress.
We were able to obtain low jitter values with a setup similar to Tabor. We have
used a high frequency generator to make 1.428 GHz at 0 dBm. We then
synchronized the frequency generator and the 120 Hz pulser with a 10 MHz
reference line. 1.428 GHz was sent onto CLK IN port and 120 Hz signal was used
to trigger the Proteus, and the scope.

Also note we changed the trigger level from 0.1 which had noticable affect on
the jitter.