Control the amplifier trigger delay
The test setup can control and take data from the amplifier and pulser. The
amplifier is powered by the RF waveform made by the AWG. The AWG is triggered
by one DG645 function generator at 120 Hz. The delay of this trigger should be
controllable remotely, because the actual time coincidence of the beam will be
slightly away from this 120 Hz trigger. I added this control in.